feat: Add views index and remove item from cart

+++ VShopWeb/Models/CartHeaderViewModel.cs
@@ -5,5 +5,5 @@ public class CartHeaderViewModel
     public int Id { get; set; }
     public string UserId { get; set; } = string.Empty;
     public string CouponCode {  get; set; } = string.Empty;
-    public double TotalAmount { get; set; } = 0.00d;
+    public decimal TotalAmount { get; set; } = 0.00m;
 }
